requirements.txt
Flask==2.0.2
gunicorn==20.1.0
python-dotenv==0.19.2
requests==2.26.0
pandas==1.3.4